Adds a design document for external compactions
In a conversation with @ctubbsii one day he asked me if a distributed queue
could be used instead of the approach proposed in #1451.  A few weeks later I
brought this up in a discussion with @billierinaldi and we worked out a
possible way to do this on a whiteboard.  This document captures that.