Fix links for staging site and other broken links
Change use of absolute paths to relative paths, so the links on the
staging site point to other pages on the staging site, rather than their
corresponding page on the production site (which may not exist yet).

Also update some other links to use https or fix broken links, if
possible.