Remove self from Reviewers lines on blog posts
Removing myself from Reviewers lines on blog posts because for me,
personally, I feel it singles me out as responsible for the content more
so than other committers. So, I would prefer not to be listed there.

In general, I don't like the Reviewers tag on blog posts for this
reason, but this change only removes myself, and does not make any
additional changes to that practice.