Add JShell version of the Accumulo Tour to website.
Added a version of the Accumulo Tour that makes use of the JShell
feature of Accumulo 2.1.x. The user will have to option to go through
the tour using the former version with the MiniAccumuloCluster or the
newer JShell feature.

Also added a new page to the tour illustrating iterators.