Use direct download links when javascript unavailable
Use closer.lua's capability to link directly to the download of a file,
rather than to another secondary download page, when initially loading
the page.

This preserves the same behavior for users when clicking on these links
if they have javascript disabled, as when javascript is enabled and
replaces these initial links with direct download links from the
selected mirror.